Reorganize the tests and add fuzzing test support
