+++ src/test/scala/com/microsoft/ml/spark/core/test/fuzzing/FuzzingTest.scala
@@ -84,32 +84,40 @@ class FuzzingTest extends TestBase {
 
   test("Verify all stages can be serialized") {
     val exemptions: Set[String] = Set(
-      "com.microsoft.ml.spark.featurize.ValueIndexerModel",
+      "com.microsoft.ml.spark.automl.BestModel",
       "com.microsoft.ml.spark.automl.TuneHyperparameters",
-      "com.microsoft.ml.spark.train.ComputePerInstanceStatistics",
-      "com.microsoft.ml.spark.cognitive.LocalNER",
-      "com.microsoft.ml.spark.cntk.train.CNTKLearner",
-      "com.microsoft.ml.spark.core.serialize.TestEstimatorBase",
-      "com.microsoft.ml.spark.featurize.DataConversion",
-      "com.microsoft.ml.spark.nn.KNNModel",
-      "com.microsoft.ml.spark.nn.ConditionalKNNModel",
-      "com.microsoft.ml.spark.train.TrainedRegressorModel",
-      "com.microsoft.ml.spark.core.serialize.MixedParamTest",
       "com.microsoft.ml.spark.automl.TuneHyperparametersModel",
-      "com.microsoft.ml.spark.lightgbm.LightGBMRegressionModel",
-      "com.microsoft.ml.spark.isolationforest.IsolationForestModel",
-      "com.microsoft.ml.spark.vw.VowpalWabbitClassificationModel",
+      "com.microsoft.ml.spark.cntk.train.CNTKLearner",
+      "com.microsoft.ml.spark.cognitive.LocalNER",
       "com.microsoft.ml.spark.core.serialize.ComplexParamTest",
-      "com.microsoft.ml.spark.vw.VowpalWabbitRegressionModel",
+      "com.microsoft.ml.spark.core.serialize.MixedParamTest",
       "com.microsoft.ml.spark.core.serialize.StandardParamTest",
-      "com.microsoft.ml.spark.vw.VowpalWabbitContextualBanditModel",
-      "com.microsoft.ml.spark.stages.ClassBalancerModel",
+      "com.microsoft.ml.spark.core.serialize.TestEstimatorBase",
+      "com.microsoft.ml.spark.explainers.ImageLIME",
+      "com.microsoft.ml.spark.explainers.ImageSHAP",
+      "com.microsoft.ml.spark.explainers.TabularLIME",
+      "com.microsoft.ml.spark.explainers.TabularSHAP",
+      "com.microsoft.ml.spark.explainers.TextLIME",
+      "com.microsoft.ml.spark.explainers.TextSHAP",
+      "com.microsoft.ml.spark.explainers.VectorLIME",
+      "com.microsoft.ml.spark.explainers.VectorSHAP",
       "com.microsoft.ml.spark.featurize.CleanMissingDataModel",
-      "com.microsoft.ml.spark.stages.TimerModel",
+      "com.microsoft.ml.spark.featurize.DataConversion",
+      "com.microsoft.ml.spark.featurize.ValueIndexerModel",
+      "com.microsoft.ml.spark.isolationforest.IsolationForestModel",
       "com.microsoft.ml.spark.lightgbm.LightGBMClassificationModel",
-      "com.microsoft.ml.spark.train.TrainedClassifierModel",
       "com.microsoft.ml.spark.lightgbm.LightGBMRankerModel",
-      "com.microsoft.ml.spark.automl.BestModel"
+      "com.microsoft.ml.spark.lightgbm.LightGBMRegressionModel",
+      "com.microsoft.ml.spark.nn.ConditionalKNNModel",
+      "com.microsoft.ml.spark.nn.KNNModel",
+      "com.microsoft.ml.spark.stages.ClassBalancerModel",
+      "com.microsoft.ml.spark.stages.TimerModel",
+      "com.microsoft.ml.spark.train.ComputePerInstanceStatistics",
+      "com.microsoft.ml.spark.train.TrainedClassifierModel",
+      "com.microsoft.ml.spark.train.TrainedRegressorModel",
+      "com.microsoft.ml.spark.vw.VowpalWabbitClassificationModel",
+      "com.microsoft.ml.spark.vw.VowpalWabbitContextualBanditModel",
+      "com.microsoft.ml.spark.vw.VowpalWabbitRegressionModel"
     )
     val applicableStages = pipelineStages.filter(t => !exemptions(t.getClass.getName))
     val applicableClasses = applicableStages.map(_.getClass.asInstanceOf[Class[_]]).toSet
